refactor(checker): split class constructor helpers
